server/requirements.txt
# fastapi==0.111.0
# uvicorn[standard]==0.30.1
# pydantic==2.7.4
# python-multipart==0.0.9

# # to be used in later steps (leave them here now)
# pdfplumber==0.11.4
# python-docx==1.1.2
# spacy==3.7.4
# rapidfuzz==3.9.6
# scikit-learn==1.4.2


fastapi==0.111.0
uvicorn==0.30.1            # no [standard] to avoid uvloop/websockets build
pydantic==2.7.4
python-multipart==0.0.9

pdfplumber==0.11.4
python-docx==1.1.2
rapidfuzz==3.9.6

# Security & production dependencies
python-magic==0.4.27  # Optional: enhanced MIME validation (requires libmagic system library)
slowapi==0.1.9
gunicorn==21.2.0

# Configuration
python-dotenv==1.0.0
